+++ Pipfile
@@ -1,0 +1,17 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+flask = "*"
+flask-pymongo = "*"
+python-dotenv = "*"
+flask-bcrypt = "*"
+mongo = {extras = ["srv"], version = "*"}
+dnspython = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.8"
