+++ source/extensions.py
@@ -1,0 +1,5 @@
+from flask_pymongo import PyMongo
+from flask_bcrypt import Bcrypt
+
+mongo = PyMongo()
+bcrypt = Bcrypt()
